Allow implicitly adding media to previous post
Store the last post each user has made. If they upload another image
without any text in the next few seconds, add that image to the post.
This fixes clients that upload multiple images as individual posts.

When sharing multiple images on iOS from an external app it shares each
one in an individual post. It also doesn't always share them in the
original order (the post with the text can come after the pictures).

To fix this, all messages are pushed into a queue. Once no messages are
received for a few seconds, the queue is processed. All the queued
messages are sorted and processed synchronously.

Fix #1