Put each attachment on a new line in the generated markdown
This works around an issue where video thumbnails (which are inserted
using HTML) would fail to be processed by Pelican when separated by a
single space. This resulted in the raw markdown appearing the HTML
instead of it being rendered.

Either removing the space or changing it to a new line resolves this
issue. This commit uses a new line since it makes it easier to read the
markdown and makes sure the markdown parser doesn't accidentally combine
attachments.