Fix issue where buttons failed to match the post state
`discord.py` v2 no longer makes edits to `discord.Message` objects
in-place when calling `message.edit()`, instead returning a new message
with the edit applied.

To fix this, we return the new message from every `cmd_reply_*` function
so the new pulished/unpublished state can be pulled out of it and used
to create the proper buttons.