Avoid blocking the event loop with long regenerates
Uses `asyncio.to_thread` to run the regeneration function in a
background thread so that the Discord bot can continue to respond to
messages in the meantime. Also, now that messages can be processed at
the same time as regeneration is occurring, added debouncing to the
regeneration function as well as a lock.